// src/generated/models/Aur_email_queuesesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Aur_email_queuesesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Aur_email_queuesesstatecode = keyof typeof Aur_email_queuesesstatecode;
export const Aur_email_queuesesstatuscode = {
  1: 'Draft',
  2: 'Inactive',
  751820001: 'NotConvertedtoInquiry',
  751820002: 'ConvertedtoInquiry'
} as const;
export type Aur_email_queuesesstatuscode = keyof typeof Aur_email_queuesesstatuscode;

export interface Aur_email_queuesesBase {
  aur_email_body?: string;
  aur_email_queuesid: string;
  "aur_inquiry@odata.bind"?: string;
  aur_mail_box?: string;
  aur_name?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Aur_email_queuesesstatecode;
  statuscode?: Aur_email_queuesesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Aur_email_queueses extends Aur_email_queuesesBase {
  aur_inquiryname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  aur_inquiry?: object;
  _aur_inquiry_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}